refactor(ui): organize ui into folders
