+++ src/ui/pages/LevelSelectPage/LevelSelectPage.jsx
@@ -1,11 +1,10 @@
 import React from 'react';
 import { css } from '@emotion/react';
-import { theme } from '../../shared/theme';
-import { LevelButton, LEVEL_BUTTON_SIZE } from '../LevelButton';
-import { PageHeader } from '../PageHeader';
-import { scenes } from '../../game/scenes';
-import { LevelScene } from '../../game/scenes/LevelScene';
-import backgroundUrl from '../../public/img/background.jpg';
+import { theme } from '../../../shared/theme';
+import { PageHeader, LevelButton, LEVEL_BUTTON_SIZE } from '../../components';
+import { scenes } from '../../../game/scenes';
+import { LevelScene } from '../../../game/scenes/LevelScene';
+import backgroundUrl from '../../../public/img/background.jpg';
 
 const LEVEL_GRID_SPACING = '2rem';
